fixed linters

+++ examples/long_task_api/tasks.py
@@ -1,6 +1,6 @@
 import asyncio
 import time
-from random import random
+from random import random  # noqa: DUO102
 
 
 async def execute_async_cpu(load_factor: int = 9) -> int:
@@ -8,13 +8,13 @@ async def execute_async_cpu(load_factor: int = 9) -> int:
 
 
 async def execute_async_io(load_factor: int = 9) -> float:
-    seconds = random() * (10 ** (load_factor - 7)) / 2
+    seconds = float(random() * (10 ** (load_factor - 7)) / 2)  # noqa: S311
     await asyncio.sleep(seconds)
     return seconds
 
 
 def execute_sync_cpu(load_factor: int = 9) -> int:
-    iterations = int(random() * (10**load_factor))
+    iterations = int(random() * (10**load_factor))  # noqa: S311
     acc = 0
     for i in range(iterations):
         acc += i
@@ -23,6 +23,6 @@ def execute_sync_cpu(load_factor: int = 9) -> int:
 
 
 def execute_sync_io(load_factor: int = 9) -> float:
-    seconds = random() * (10 ** (load_factor - 7)) / 2
+    seconds = float(random() * (10 ** (load_factor - 7)) / 2)  # noqa: S311
     time.sleep(seconds)
     return seconds
